binutils: Drop mips XLP support patch
The kernel has dropped this as of 5.16 and we don't want to carry such patches
without active maintainers for such targets.

It isn't clear who would even have such hardware and it isn't something we can
support. It would be best maintained as a separate layer by those who can test
it if needed.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>