apr: remove obsolete support for renamed libtool
libtool is now longer renamed to ${host}-libtool, so remove the changes
to support this.

This means that apr now installs libtool into the build-1 folder, but
as this has never been needed before (as we use the system libtool) we
can remove it (it contains build paths so is unreproducible).  Also add
a RDEPENDS on libtool for the target -dev package.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>