manpages: Fix override/append ordering and hence task signatures
The append/override ordering was meaning that whilst the data was being picked
up later in the package processing, it wasn't being picked up by do_package.
This means changes to qemu options were not causing task signatures to change.
Fix the variable ordering with the append to correct this.

The whitespace fix is deliberate to cause output to change and hopefully avoid
hash invalidation issues from unchanged output.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>