meson: Handle qemu riscv issues
In qemu, they merged the 32 and 64 bit riscv variants into one target for meson
purposes but this isn't officially supported by upstream meson. In normal builds
this would just be a warning but in OE, this is fatal.

Avoid the fatal errors for now and build as qemu intended whilst this is resolved
by the upstreams.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>