package_deb/ipk/rpm: Add more minimal do_build dependencies back
The dependencies for do_build became a little too minimal after the
removal of recrdeptask since "bitbake go" would not package go-runtime
despite it being in DEPENDS and the resulting package having a dependency
on it.

A reasonable compromise is probably rdeptask instead of recrdeptask
which is a lot lower overhead but makes the build target more useful
and importantly, lets world builds do what you'd expect them to.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>