libsdl2: Fix X11 configure options
We're seeing various failures where the X11 headers are found on the native
system but not all libraries are present and hence autoconfiguration of the
X11 subcomponents fails.

We don't list any of these X11 subcomponents as a dependency so disable
them by default. Configuration and dependencies can be added if people need
them.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>