curl: Rework openssl and random PACKAGECONFIGs
The existing ssl PACKAGECONFIG makes it slightly annoying to use another
crypto provider while removing openssl. Since --with-ssl is just a
deprecated version of --with-openssl, rename the PACKAGECONFIG to use
the newer preferred name. Note that --without-ssl implies no crypto
provider at all, and should only be used when trying to disable all
crypto support.

Move --with-random to it's own option, since it is useful for other
crypto providers, not just openssl.

Signed-off-by: Robert Joslyn <robert.joslyn@redrectangle.org>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>