abi_version: Bump HASHEQUIV_HASH_VERSION
After the changes to the native outhash files to ensure populate_sysroot for
native recipes correctly accounts for runtime dependencies, this doesn't reset
the target output namespace and means reproducibility issues are still triggering
on our test infrastructure.

Bump the version being used for hashequiv to ensure we avoid these issues with
a new namespace.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>