python3-pip: upgrade 22.0.2 -> 22.0.3
Changelog:
=========
Features
-------
Print the exception via rich.traceback, when running with --debug. (#10791)

Bug Fixes
---------
Only calculate topological installation order, for packages that are going to
be installed/upgraded.

This fixes an AssertionError that occured when determining installation order,
for a very specific combination of upgrading-already-installed-package + change
of dependencies + fetching some packages from a package index. This combination
was especially common in Read the Docs' builds. (#10851)

Use html.parser by default, instead of falling back to html5lib when
--use-deprecated=html5lib is not passed. (#10869)

Improved Documentation
----------------------
Clarify that using per-requirement overrides disables the usage of wheels. (#9674)

Signed-off-by: Wang Mingyu <wangmy@fujitsu.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>