selftest/bbtests: Add tests for git floating tag resolution
This test ensures that where a floating tag is used for a git url, there
aren't a number of different bugs such as inconsistent source revisions
for the tag within a given build and that the recipe is reparsed at each
run to ensure the tag is checked.

A test is also added to ensure that the build fails if SRCPV (get_srcrev())
isn't used yet there is a floating tag.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>