weston-init: disable systemd watchdog option
There is no other unit in oe-core where watchdog is enabled, and
this feature starts to behave weirdly when test_date is adjusting the clock,
which means weston is restarted without need, so disable for weston
test stability.

The issue won't show up on kvm qemu (x86/x86_64) because it runs
too quickly, but non-kvm qemu (ppc/mips/arm) all fail in the same
way due to slower timings.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>