libsndfile1: bump to version 1.0.31
It seems that the homepage has moved (for a while now) from
  http://www.mega-nerd.com/libsndfile/
to
  https://libsndfile.github.io/libsndfile/

On Github, the development group seems to include Erik de Castro Lopo (the
original author of libsndfile1).
Link: https://github.com/orgs/libsndfile/people

All backported CVE patches are in version 1.0.31 (I've checked).

The supported format page [1], mentions that Ogg/Opus is supported since
1.0.29, however that isn't currently added in this patch. And it may
require libopus.
And mp3 is supported at around version 1.1.0, but that version is in beta.

This change focuses solely on updating to 1.0.31 and removing all
backported patches.

[1] https://libsndfile.github.io/libsndfile/formats.html

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>