python3-hypothesis: upgrade 6.37.2 -> 6.39.0
Changelog:
=========
This release improves Hypothesis' handling of positional-only arguments,
which are now allowed @st.composite strategies.

On Python 3.8 and later, the first arguments to builds() and from_model()
are now natively positional-only. In cases which were already errors, the
TypeError from incorrect usage will therefore be raises immediately when
the function is called, rather than when the strategy object is used.

Signed-off-by: Wang Mingyu <wangmy@fujitsu.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>