python3-native: Drop opt-1 and opt-2 pyc files
There are over 3,000 of these in python3-native (of 8,000+ files total) and
copying them to sysroots all the time seems pointless, particularly since
they're only used if python is run with the -O or -OO parameters.

Get rid of them and save the overhead. This is particularly pronounced in
builds using the api-documentation distro feature.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>